by ljc

+++ ljc-order/src/main/java/com/ljc/order/vo/OrderConfirmVo.java
@@ -23,7 +23,7 @@ public class OrderConfirmVo {
     /**
      * 优惠券（会员积分）
      **/
-    private Integer integration;
+    private Long integration;
 
     /**
      * 防止重复提交的令牌
