+++ ljc-resource-server/src/main/java/com/ljc/resource/server/security/SecurityUtil.java
@@ -26,7 +26,7 @@ public class SecurityUtil {
         return getUser() == null ? null : getUser().getClientId();
     }
 
-    public static Integer getIntegration() {
+    public static Long getIntegration() {
         return getUser() == null ? null : getUser().getAttribute("integration");
     }
